// whoxa_frontend/src/types/ResType.ts
export interface BlockUserRes {
  is_block: boolean;
  success: boolean;
  message: string;
}